backend/requirements.txt
uvicorn
python-dotenv
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
chromadb
python-multipart
pypdf
